Split Sort test code into three files.
