+++ src/System.Memory/tests/System.Memory.Tests.csproj
@@ -9,7 +9,6 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'netstandard-Release|AnyCPU'" />
   <ItemGroup>
     <Compile Include="AllocationHelper.cs" />
-    <Compile Include="Span\Sort.Fillers.cs" />
     <Compile Include="TInt.cs" />
     <Compile Include="TestException.cs" />
     <Compile Include="TestHelpers.cs" />
@@ -54,6 +53,8 @@
     <Compile Include="Span\SequenceEqual.char.cs" />
     <Compile Include="Span\Slice.cs" />
     <Compile Include="Span\Sort.cs" />
+    <Compile Include="Span\Sort.Fillers.cs" />
+    <Compile Include="Span\Sort.Helpers.cs" />
     <Compile Include="Span\StartsWith.T.cs" />
     <Compile Include="Span\StartsWith.byte.cs" />
     <Compile Include="Span\ToArray.cs" />
